+++ CHANGELOG.md
@@ -162,3 +162,4 @@
 * Define `Constant` data type
 * Define `point` expression
 * Define `TT_T_I` data type
+* Define `U_UU_UUU_V_I_II_III_UUUU` data type